requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
imageio-ffmpeg>=0.4.10
yt-dlp>=2025.10.22
openai-whisper>=2023.9.0
torch>=2.2.0,<3.0